fix license and __all__. test=develop
