+++ python/paddle/fluid/dataloader/worker.py
@@ -24,7 +24,6 @@ from .fetcher import _IterableDatasetFetcher, _MapDatasetFetcher
 from ..multiprocess_utils import _cleanup_mmap, CleanupFuncRegistrar, MP_STATUS_CHECK_INTERVAL
 from ..framework import in_dygraph_mode
 from .flat import _flatten_batch
-from .state import _generate_states
 
 # NOTE: queue has a different name in python2 and python3
 if six.PY2:
@@ -167,6 +166,89 @@ class _WorkerException(object):
         if getattr(self.exc_type, "message", None):
             raise self.exc_type(message=msg)
         raise self.exc_type(msg)
+
+
+# The function `_generate_states` is adapted from `numpy.random.SeedSequence`
+# from https://github.com/numpy/numpy/blob/main/numpy/random/bit_generator.pyx
+# Here is the copyright:
+
+# SeedSequence is derived from Melissa E. O'Neill's C++11 `std::seed_seq`
+# implementation, as it has a lot of nice properties that we want.
+# https://gist.github.com/imneme/540829265469e673d045
+# http://www.pcg-random.org/posts/developing-a-seed_seq-alternative.html
+
+# The MIT License (MIT)
+
+# Copyright (c) 2015 Melissa E. O'Neill
+# Copyright (c) 2019 NumPy Developers
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+INIT_A = 0x43b0d7e5
+MULT_A = 0x931e8875
+INIT_B = 0x8b51f9dd
+MULT_B = 0x58f38ded
+MIX_MULT_L = 0xca01f9dd
+MIX_MULT_R = 0x4973f715
+XSHIFT = np.dtype(np.uint32).itemsize * 8 // 2
+MASK32 = 0xFFFFFFFF
+
+
+def _generate_states(base_seed=0, worker_id=0):
+    # init hash constant
+    hash_const_A = INIT_A
+    hash_const_B = INIT_B
+
+    def hash(value):
+        nonlocal hash_const_A
+        value = (value ^ hash_const_A) & MASK32
+        hash_const_A = (hash_const_A * MULT_A) & MASK32
+        value = (value * hash_const_A) & MASK32
+        value = (value ^ (value >> XSHIFT)) & MASK32
+        return value
+
+    def mix(x, y):
+        result_x = (MIX_MULT_L * x) & MASK32
+        result_y = (MIX_MULT_R * y) & MASK32
+        result = (result_x - result_y) & MASK32
+        result = (result ^ (result >> XSHIFT)) & MASK32
+        return result
+
+    # init entropys with based_seed and worker_id and calculate pool
+    entropys = [worker_id, base_seed & MASK32, base_seed >> 32, 0]
+    pool = [hash(entropy) for entropy in entropys]
+
+    # mix all bits together
+    for i in range(len(pool)):
+        for j in range(len(pool)):
+            if i != j:
+                pool[j] = mix(pool[j], hash(pool[i]))
+
+    states = []
+    for p in pool:
+        state = (p ^ hash_const_B) & MASK32
+        hash_const_B = (hash_const_B * MULT_B) & MASK32
+        state = (state * hash_const_B) & MASK32
+        state = (state ^ (state >> XSHIFT)) & MASK32
+        states.append(state)
+
+    return states
 
 
 def _worker_loop(dataset, dataset_kind, indices_queue, out_queue, done_event,
